“第二轮”

+++ src/main/java/com/bank/reconciliation/entity/BankOrder.java
@@ -23,7 +23,7 @@ public class BankOrder {
     @Column(length = 64)
     private String unionpayTraceNo;
 
-    @Column(nullable = false, precision = 18, scale = 2)
+    @Column(nullable = false, precision = 32, scale = 2)
     private BigDecimal amount;
 
     @Column(length = 32)
